chore: Restyling transaction summary screen
