+++ .env.production
@@ -65,8 +65,6 @@ PREMIUM_MESSAGES_OPT_IN_ENABLED=YES
 CDC_ENABLED = YES
 # Scan additional barcodes (E.g. Data Matrix) in the payment section
 SCAN_ADDITIONAL_BARCODES_ENABLED = YES
-# Redesign of the transaction summary screen
-NEW_TRANSACTION_SUMMARY_ENABLED=YES
 # Opt-in for reminder push notifications
 REMINDERS_OPT_IN_ENABLED=YES
 # FCI (Firma con IO) feature
