+++ locales/it/index.yml
@@ -1406,6 +1406,7 @@ wallet:
     amount: Importo
     total: Totale pagato
     expireDate: Scadenza
+    dueDate: Entro il
     tranche: Rata
     info: "Per informazioni:"
     tel: "Tel:"
