+++ ts/features/pn/analytics/index.ts
@@ -2,7 +2,7 @@ import { pipe } from "fp-ts/lib/function";
 import * as A from "fp-ts/lib/Array";
 import * as O from "fp-ts/lib/Option";
 import { mixpanelTrack } from "../../../mixpanel";
-import { TransactionSummaryErrorContent } from "../../../screens/wallet/payment/NewTransactionSummaryScreen";
+import { TransactionSummaryErrorContent } from "../../../screens/wallet/payment/TransactionSummaryScreen";
 import { PNMessage } from "../../pn/store/types/types";
 import { NotificationStatusHistoryElement } from "../../../../definitions/pn/NotificationStatusHistoryElement";
 import { UIAttachment } from "../../../store/reducers/entities/messages/types";
