+++ ts/features/pn/components/PnMessagePayment.tsx
@@ -10,7 +10,7 @@ import { TransactionSummary } from "../../../screens/wallet/payment/components/T
 import { TransactionSummaryErrorDetails } from "../../../screens/wallet/payment/components/TransactionSummaryErrorDetails";
 import { PotFromActions } from "../../../types/utils";
 import { paymentVerifica } from "../../../store/actions/wallet/payment";
-import { TransactionSummaryError } from "../../../screens/wallet/payment/NewTransactionSummaryScreen";
+import { TransactionSummaryError } from "../../../screens/wallet/payment/TransactionSummaryScreen";
 import { UIMessageId } from "../../../store/reducers/entities/messages/types";
 import { InfoBox } from "../../../components/box/InfoBox";
 import { navigateToPnCancelledMessagePaidPaymentScreen } from "../navigation/actions";
