+++ ts/screens/wallet/payment/TransactionSummaryScreen.tsx
@@ -1,42 +1,31 @@
-import {
-  AmountInEuroCents,
-  PaymentNoticeNumberFromString,
-  RptId
-} from "@pagopa/io-pagopa-commons/lib/pagopa";
+import { PaymentNoticeNumberFromString } from "@pagopa/io-pagopa-commons/lib/pagopa";
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import { pipe } from "fp-ts/lib/function";
 import * as O from "fp-ts/lib/Option";
 import { ActionSheet } from "native-base";
-import * as React from "react";
-import { View, SafeAreaView } from "react-native";
+import React, { useEffect } from "react";
+import { SafeAreaView, ScrollView, StyleSheet } from "react-native";
+import { ButtonSolid, ContentWrapper } from "@pagopa/io-app-design-system";
 import { connect } from "react-redux";
-
-import { IOColors, VSpacer } from "@pagopa/io-app-design-system";
 import { PaymentRequestsGetResponse } from "../../../../definitions/backend/PaymentRequestsGetResponse";
-import { Body } from "../../../components/core/typography/Body";
-import { H2 } from "../../../components/core/typography/H2";
 import { IOStyles } from "../../../components/core/variables/IOStyles";
-import { withLoadingSpinner } from "../../../components/helpers/withLoadingSpinner";
-import ItemSeparatorComponent from "../../../components/ItemSeparatorComponent";
 import BaseScreenComponent from "../../../components/screens/BaseScreenComponent";
-import FocusAwareStatusBar from "../../../components/ui/FocusAwareStatusBar";
-import FooterWithButtons from "../../../components/ui/FooterWithButtons";
-import { PaymentSummaryComponent } from "../../../components/wallet/PaymentSummaryComponent";
-import { SlidedContentComponent } from "../../../components/wallet/SlidedContentComponent";
 import {
   isError,
   isLoading as isRemoteLoading,
   isUndefined
 } from "../../../features/bonus/bpd/model/RemoteValue";
+import {
+  zendeskSelectedCategory,
+  zendeskSupportStart
+} from "../../../features/zendesk/store/actions";
 import I18n from "../../../i18n";
 import { IOStackNavigationRouteProps } from "../../../navigation/params/AppParamsList";
 import { WalletParamsList } from "../../../navigation/params/WalletParamsList";
 import {
-  navigateToPaymentManualDataInsertion,
   navigateToPaymentPickPaymentMethodScreen,
   navigateToPaymentTransactionErrorScreen,
-  navigateToWalletAddPaymentMethod,
-  navigateToWalletHome
+  navigateToWalletAddPaymentMethod
 } from "../../../store/actions/navigation";
 import { Dispatch } from "../../../store/actions/types";
 import {
@@ -46,7 +35,6 @@ import {
   paymentCompletedSuccess,
   paymentIdPolling,
   paymentInitializeState,
-  PaymentStartOrigin,
   paymentVerifica,
   runDeleteActivePaymentSaga,
   runStartOrResumePaymentActivationSaga
@@ -61,23 +49,93 @@ import {
   getFavoriteWallet,
   withPaymentFeatureSelector
 } from "../../../store/reducers/wallet/wallets";
+import customVariables from "../../../theme/variables";
 import { PayloadForAction } from "../../../types/utils";
-import { cleanTransactionDescription } from "../../../utils/payment";
+import { emptyContextualHelp } from "../../../utils/emptyContextualHelp";
+import { useOnFirstRender } from "../../../utils/hooks/useOnFirstRender";
+import {
+  DetailV2Keys,
+  getCodiceAvviso,
+  getV2ErrorMainType,
+  isDuplicatedPayment
+} from "../../../utils/payment";
 import { alertNoPayablePaymentMethods } from "../../../utils/paymentMethod";
 import { showToast } from "../../../utils/showToast";
 import {
-  centsToAmount,
-  formatNumberAmount
-} from "../../../utils/stringBuilder";
-import { formatTextRecipient } from "../../../utils/strings";
+  addTicketCustomField,
+  appendLog,
+  resetCustomFields,
+  zendeskCategoryId,
+  zendeskPaymentCategory,
+  zendeskPaymentFailure,
+  zendeskPaymentNav,
+  zendeskPaymentOrgFiscalCode,
+  zendeskPaymentStartOrigin
+} from "../../../utils/supportAssistance";
 import { dispatchPickPspOrConfirm } from "./common";
+import { TransactionSummary } from "./components/TransactionSummary";
+import { TransactionSummaryErrorDetails } from "./components/TransactionSummaryErrorDetails";
+import { TransactionSummaryStatus } from "./components/TransactionSummaryStatus";
 
-export type TransactionSummaryScreenNavigationParams = Readonly<{
-  rptId: RptId;
-  initialAmount: AmountInEuroCents;
-  paymentStartOrigin: PaymentStartOrigin;
-  messageId?: string;
-}>;
+export type TransactionSummaryErrorContent = PayloadForAction<
+  | typeof paymentVerifica["failure"]
+  | typeof paymentAttiva["failure"]
+  | typeof paymentIdPolling["failure"]
+>;
+
+export type TransactionSummaryError = O.Option<TransactionSummaryErrorContent>;
+
+const styles = StyleSheet.create({
+  container: {
+    paddingHorizontal: customVariables.contentPadding
+  }
+});
+
+const renderFooter = (
+  isLoading: boolean,
+  error: TransactionSummaryError,
+  continueWithPayment: () => void,
+  help: () => void
+) => {
+  if (O.isSome(error)) {
+    const errorOrUndefined = O.toUndefined(error);
+    const errorType = getV2ErrorMainType(errorOrUndefined as DetailV2Keys);
+    switch (errorType) {
+      case "TECHNICAL":
+      case "DATA":
+      case "UNCOVERED":
+        return (
+          <ButtonSolid
+            onPress={help}
+            fullWidth
+            accessibilityLabel={I18n.t("payment.details.info.buttons.help")}
+            label={I18n.t("payment.details.info.buttons.help")}
+          />
+        );
+      // There's a strange case where error is 'some' but
+      // its value is undefined (e.g. network error).
+      // In this case we fallback to the 'continue' CTA
+      // so that the user can eventually retry.
+      case undefined:
+        break;
+      default:
+        return <></>;
+    }
+  }
+  if (isLoading) {
+    // TODO: Add a ButtonOutline with a loading spinner as soon as it's available by DS
+    return <></>;
+  }
+
+  return (
+    <ButtonSolid
+      fullWidth
+      onPress={continueWithPayment}
+      label={I18n.t("wallet.continue")}
+      accessibilityLabel={I18n.t("wallet.continue")}
+    />
+  );
+};
 
 type OwnProps = IOStackNavigationRouteProps<
   WalletParamsList,
@@ -88,53 +146,72 @@ type Props = ReturnType<typeof mapStateToProps> &
   ReturnType<typeof mapDispatchToProps> &
   OwnProps;
 
-/**
- * This screen shows the transaction details once the payment has been verified
- * (to check if it is valid and if the amount has been updated)
- */
-class TransactionSummaryScreen extends React.Component<Props> {
-  public componentDidMount() {
-    if (pot.isNone(this.props.potVerifica)) {
-      // on component mount, fetch the payment summary if we haven't already
-      this.props.dispatchPaymentVerificaRequest();
+const TransactionSummaryScreen = ({
+  isLoading,
+  error,
+  paymentVerification,
+  verifyPayment,
+  onDuplicatedPayment,
+  navigateToPaymentTransactionError,
+  walletById,
+  loadWallets,
+  navigation,
+  route,
+  continueWithPayment,
+  maybeFavoriteWallet,
+  hasPayableMethods,
+  paymentId,
+  backToEntrypointPayment,
+  resetPayment,
+  startAssistanceRequest
+}: Props): React.ReactElement => {
+  useOnFirstRender(() => {
+    if (pot.isNone(paymentVerification)) {
+      verifyPayment();
     }
-    if (!pot.isSome(this.props.walletById)) {
-      this.props.loadWallets();
+    if (!pot.isSome(walletById)) {
+      loadWallets();
     }
-  }
+  });
 
-  public componentDidUpdate(prevProps: Props) {
-    const { error, potVerifica } = this.props;
-    if (O.toUndefined(error) !== O.toUndefined(prevProps.error)) {
-      // in case the verifica returns an error indicating the payment has been
-      // already completed for this notice, we update the payment state so that
-      // the notice result paid
-      pipe(
-        error,
-        O.filter(_ => _ === "PAA_PAGAMENTO_DUPLICATO"),
-        O.map(_ => this.props.onDuplicatedPayment())
-      );
-      if (O.isSome(error)) {
-        this.props.navigateToPaymentTransactionError(error);
-      }
-    } else if (
-      potVerifica !== prevProps.potVerifica &&
-      pot.isError(potVerifica)
-    ) {
-      this.props.navigateToPaymentTransactionError(O.some(potVerifica.error));
-    } else if (
-      // this is the case when the component is already mounted (eg. process more payments)
-      // we check if the rptId is different from the previous one, in that case fire the dispatchPaymentVerificaRequest
-      pot.isNone(this.props.potVerifica) &&
-      this.props.route.params.rptId.paymentNoticeNumber !==
-        prevProps.route.params.rptId.paymentNoticeNumber
-    ) {
-      this.props.dispatchPaymentVerificaRequest();
-    }
-  }
+  // We show inline error status only if the payment starts
+  // from a message and the verification fails. In all the other
+  // cases we present the fullscreen error message.
+  const paymentStartOrigin = route.params.paymentStartOrigin;
+  const showsInlineError = paymentStartOrigin === "message";
 
-  private handleBackPress = () => {
-    if (pot.isSome(this.props.paymentId)) {
+  const errorOrUndefined = O.toUndefined(error);
+  const isError = O.isSome(error);
+
+  const isPaid = isDuplicatedPayment(error);
+  useEffect(() => {
+    if (!isError) {
+      return;
+    }
+    if (isPaid) {
+      onDuplicatedPayment();
+    }
+    // in case of a payment verification error we should navigate
+    // to the error screen only if showsInlineError is false
+    // in any other case we should always navigate to the error screen
+    if (
+      (pot.isError(paymentVerification) && !showsInlineError) ||
+      (!pot.isError(paymentVerification) && isError)
+    ) {
+      navigateToPaymentTransactionError(O.fromNullable(errorOrUndefined));
+    }
+  }, [
+    isError,
+    errorOrUndefined,
+    onDuplicatedPayment,
+    navigateToPaymentTransactionError,
+    showsInlineError,
+    paymentVerification,
+    isPaid
+  ]);
+
+  const goBack = () => {
+    if (pot.isSome(paymentId)) {
       // If we have a paymentId (payment check already done) we need to
       // ask the user to cancel the payment and in case reset it
       ActionSheet.show(
@@ -149,8 +226,8 @@ class TransactionSummaryScreen extends React.Component<Props> {
         },
         buttonIndex => {
           if (buttonIndex === 0) {
-            this.props.backToEntrypointPayment();
-            this.props.resetPayment();
+            backToEntrypointPayment();
+            resetPayment();
             showToast(
               I18n.t("wallet.ConfirmPayment.cancelPaymentSuccess"),
               "success"
@@ -159,210 +236,95 @@ class TransactionSummaryScreen extends React.Component<Props> {
         }
       );
     } else {
-      this.props.navigation.goBack();
+      navigation.goBack();
     }
   };
 
-  private getSecondaryButtonProps = () => ({
-    bordered: pot.isNone(this.props.paymentId),
-    cancel: pot.isSome(this.props.paymentId),
-    onPress: this.handleBackPress,
-    title: I18n.t("global.buttons.cancel")
-  });
+  const rptId = route.params.rptId;
+  const messageId = route.params.messageId;
 
-  private renderFooterSingleButton() {
-    return (
-      <FooterWithButtons
-        type={"SingleButton"}
-        leftButton={this.getSecondaryButtonProps()}
-      />
-    );
-  }
+  const paymentNoticeNumber = PaymentNoticeNumberFromString.encode(
+    rptId.paymentNoticeNumber
+  );
 
-  private handleContinueOnPress = (verifica: PaymentRequestsGetResponse) => {
-    const { maybeFavoriteWallet, hasPayableMethods } = this.props;
-    if (hasPayableMethods) {
-      this.props.startOrResumePayment(
-        verifica,
-        maybeFavoriteWallet,
-        hasPayableMethods
-      );
-      return;
-    }
-    alertNoPayablePaymentMethods(this.props.navigateToWalletAddPaymentMethod);
-  };
+  /**
+   * try to show the fiscal code coming from the 'verification' API
+   * otherwise (it could be an issue with the API) use the rptID coming from
+   * static data (e.g. message, qrcode, manual insertion, etc.)
+   */
+  const organizationFiscalCode = pipe(
+    pot.toOption(paymentVerification),
+    O.chainNullableK(
+      _ => _.enteBeneficiario?.identificativoUnivocoBeneficiario
+    ),
+    O.getOrElse(() => rptId.organizationFiscalCode)
+  );
 
-  private renderFooterButtons() {
-    const { potVerifica } = this.props;
-    const basePrimaryButtonProps = {
-      primary: true,
-      title: I18n.t("wallet.continue")
-    };
-
-    const primaryButtonProps =
-      pot.isSome(potVerifica) &&
-      !(pot.isLoading(potVerifica) || pot.isError(potVerifica))
-        ? {
-            ...basePrimaryButtonProps,
-            disabled: false,
-            onPress: () => this.handleContinueOnPress(potVerifica.value)
-          }
-        : {
-            ...basePrimaryButtonProps,
-            disabled: true
-          };
-
-    return (
-      <FooterWithButtons
-        type={"TwoButtonsInlineThird"}
-        leftButton={this.getSecondaryButtonProps()}
-        rightButton={primaryButtonProps}
-      />
-    );
-  }
-
-  private getFooterButtons = () =>
-    pipe(
-      this.props.error,
-      O.fold(
-        () => this.renderFooterButtons(),
-        error =>
-          error === "PAA_PAGAMENTO_DUPLICATO"
-            ? this.renderFooterSingleButton()
-            : this.renderFooterButtons()
-      )
-    );
-
-  public render(): React.ReactNode {
-    const rptId: RptId = this.props.route.params.rptId;
-
-    const { potVerifica } = this.props;
-
-    const recipient = pipe(
-      pot.toOption(potVerifica),
-      O.chainNullableK(_ => _.enteBeneficiario),
-      O.map(formatTextRecipient)
-    );
-
-    /**
-     * try to show the organization fiscal code coming from the 'verifica' API
-     * otherwise (it could be an issue with the API) it fallbacks on rptID coming from
-     * static data: message, qrcode, manual insertion
-     */
-    const organizationFiscalCode: string = pipe(
-      pot.toOption(potVerifica),
-      O.chainNullableK(
-        _ => _.enteBeneficiario?.identificativoUnivocoBeneficiario
-      ),
-      O.getOrElse(() => rptId.organizationFiscalCode)
-    );
-
-    const currentAmount: string = pot.getOrElse(
-      pot.map(potVerifica, (verifica: PaymentRequestsGetResponse) =>
-        formatNumberAmount(
-          centsToAmount(verifica.importoSingoloVersamento),
-          true
-        )
-      ),
-      "-"
-    );
-
-    const transactionDescription = pot.getOrElse<string>(
-      pot.mapNullable(potVerifica, pv =>
-        cleanTransactionDescription(pv.causaleVersamento)
-      ),
-      "-"
-    );
-
-    const standardRow = (label: string, value: string) => (
-      <View style={IOStyles.rowSpaceBetween}>
-        <Body color="white">{label}</Body>
-        <Body weight="SemiBold" color="white">
-          {value}
-        </Body>
-      </View>
-    );
-
-    return (
-      <BaseScreenComponent
-        goBack={this.handleBackPress}
-        dark={true}
-        headerBackgroundColor={IOColors.milderGray}
-      >
-        <FocusAwareStatusBar
-          backgroundColor={IOColors.milderGray}
-          barStyle={"light-content"}
-        />
-        <SafeAreaView style={IOStyles.flex}>
-          <SlidedContentComponent dark={true}>
-            <PaymentSummaryComponent
-              dark={true}
-              title={I18n.t("wallet.firstTransactionSummary.title")}
-              description={transactionDescription}
-              recipient={pipe(
-                recipient,
-                O.fold(
-                  () => "-",
-                  r => r
-                )
-              )}
+  return (
+    <BaseScreenComponent
+      backButtonTestID={"back-button-transaction-summary"}
+      goBack={goBack}
+      contextualHelp={emptyContextualHelp}
+      headerTitle={I18n.t("wallet.ConfirmPayment.paymentInformations")}
+    >
+      <SafeAreaView style={IOStyles.flex}>
+        {showsInlineError && <TransactionSummaryStatus error={error} />}
+        <ScrollView style={styles.container}>
+          <TransactionSummary
+            paymentVerification={paymentVerification}
+            paymentNoticeNumber={paymentNoticeNumber}
+            organizationFiscalCode={organizationFiscalCode}
+            isPaid={isPaid}
+          />
+          {showsInlineError && pot.isError(paymentVerification) && (
+            <TransactionSummaryErrorDetails
+              error={error}
+              paymentNoticeNumber={paymentNoticeNumber}
+              organizationFiscalCode={organizationFiscalCode}
+              messageId={messageId}
             />
-
-            <VSpacer size={24} />
-            <ItemSeparatorComponent noPadded={true} />
-            <VSpacer size={24} />
-
-            {/** Amount to pay */}
-            <View style={IOStyles.rowSpaceBetween}>
-              <View style={IOStyles.rowSpaceBetween}>
-                <H2 color="white" weight="SemiBold">
-                  {I18n.t("wallet.firstTransactionSummary.updatedAmount")}
-                </H2>
-              </View>
-              <H2 color="white" weight="Bold">
-                {currentAmount}
-              </H2>
-            </View>
-
-            <React.Fragment>
-              <VSpacer size={8} />
-              <Body color="white">
-                {I18n.t("wallet.firstTransactionSummary.amountInfo.message")}
-              </Body>
-            </React.Fragment>
-            <VSpacer size={24} />
-
-            <ItemSeparatorComponent noPadded={true} />
-            <VSpacer size={24} />
-
-            {standardRow(
-              I18n.t("wallet.firstTransactionSummary.entityCode"),
-              organizationFiscalCode
-            )}
-            <VSpacer size={8} />
-            {standardRow(
-              I18n.t("payment.noticeCode"),
-              PaymentNoticeNumberFromString.encode(rptId.paymentNoticeNumber)
-            )}
-            <VSpacer size={24} />
-          </SlidedContentComponent>
-          {this.getFooterButtons()}
-        </SafeAreaView>
-      </BaseScreenComponent>
-    );
-  }
-}
+          )}
+        </ScrollView>
+        <ContentWrapper>
+          {renderFooter(
+            isLoading,
+            error,
+            () =>
+              continueWithPayment(
+                paymentVerification,
+                maybeFavoriteWallet,
+                hasPayableMethods
+              ),
+            () => startAssistanceRequest(error, messageId)
+          )}
+        </ContentWrapper>
+      </SafeAreaView>
+    </BaseScreenComponent>
+  );
+};
 
 // eslint-disable-next-line complexity,sonarjs/cognitive-complexity
 const mapStateToProps = (state: GlobalState) => {
   const { verifica, attiva, paymentId, check, pspsV2 } = state.wallet.payment;
+
+  const error: TransactionSummaryError = pot.isError(verifica)
+    ? O.some(verifica.error)
+    : pot.isError(attiva)
+    ? O.some(attiva.error)
+    : pot.isError(paymentId)
+    ? O.some(paymentId.error)
+    : pot.isError(check) || isError(pspsV2.psps)
+    ? O.some(undefined)
+    : O.none;
+
   const walletById = state.wallet.wallets.walletById;
+
   const isPaypalEnabled = isPaypalEnabledSelector(state);
   const isBPayPaymentEnabled = bancomatPayConfigSelector(state).payment;
   const favouriteWallet = pot.toUndefined(getFavoriteWallet(state));
   /**
    * the favourite will be undefined if one of these condition is true
-   * - payment method is Paypal & the relative feature flag is not enabled
+   * - payment method is PayPal & the relative feature flag is not enabled
    * - payment method is BPay & the relative feature flag is not enabled
    */
   const maybeFavoriteWallet = pipe(
@@ -379,33 +341,13 @@ const mapStateToProps = (state: GlobalState) => {
       }
     })
   );
-  const psps = pspsV2.psps;
-  const error: O.Option<
-    PayloadForAction<
-      | typeof paymentVerifica["failure"]
-      | typeof paymentAttiva["failure"]
-      | typeof paymentIdPolling["failure"]
-    >
-  > = pot.isError(verifica)
-    ? O.some(verifica.error)
-    : pot.isError(attiva)
-    ? O.some(attiva.error)
-    : pot.isError(paymentId)
-    ? O.some(paymentId.error)
-    : pot.isError(check) || isError(psps)
-    ? O.some(undefined)
-    : O.none;
 
-  // we need to show the spinner when the data is in the loading state
-  // and also while the logic is processing one step's response and
-  // starting the next step's loading request
-  const isLoadingVerifica = pot.isLoading(verifica);
-  const isLoadingAttiva = pot.isLoading(attiva);
-  const isLoadingWalletById = pot.isLoading(walletById);
+  const hasPayableMethods = withPaymentFeatureSelector(state).length > 0;
+
   const isLoading =
-    isLoadingWalletById ||
-    isLoadingVerifica ||
-    isLoadingAttiva ||
+    pot.isLoading(walletById) ||
+    pot.isLoading(verifica) ||
+    pot.isLoading(attiva) ||
     (O.isNone(error) && pot.isSome(attiva) && pot.isNone(paymentId)) ||
     pot.isLoading(paymentId) ||
     (O.isNone(error) && pot.isSome(paymentId) && pot.isNone(check)) ||
@@ -413,53 +355,51 @@ const mapStateToProps = (state: GlobalState) => {
     (O.isSome(maybeFavoriteWallet) &&
       O.isNone(error) &&
       pot.isSome(check) &&
-      isUndefined(psps)) ||
-    (O.isSome(maybeFavoriteWallet) && isRemoteLoading(psps));
+      isUndefined(pspsV2.psps)) ||
+    (O.isSome(maybeFavoriteWallet) && isRemoteLoading(pspsV2.psps));
 
-  const loadingCaption = isLoadingVerifica
-    ? I18n.t("wallet.firstTransactionSummary.loadingMessage.verification")
-    : isLoadingAttiva
-    ? I18n.t("wallet.firstTransactionSummary.loadingMessage.activation")
-    : isLoadingWalletById
-    ? I18n.t("wallet.firstTransactionSummary.loadingMessage.wallet")
-    : I18n.t("wallet.firstTransactionSummary.loadingMessage.generic");
-
-  const hasPayableMethods = withPaymentFeatureSelector(state).length > 0;
   return {
-    error,
+    paymentVerification: verifica,
     isLoading,
-    loadingCaption,
-    loadingOpacity: 0.98,
-    potVerifica: verifica,
-    paymentId,
+    error,
+    walletById,
     maybeFavoriteWallet,
     hasPayableMethods,
-    walletById
+    paymentId
   };
 };
 
 const mapDispatchToProps = (dispatch: Dispatch, props: OwnProps) => {
   const rptId = props.route.params.rptId;
-  const initialAmount = props.route.params.initialAmount;
   const paymentStartOrigin = props.route.params.paymentStartOrigin;
-  const isManualPaymentInsertion = paymentStartOrigin === "manual_insertion";
+  const initialAmount = props.route.params.initialAmount;
 
-  const dispatchPaymentVerificaRequest = () =>
+  const verifyPayment = () =>
     dispatch(
       paymentVerifica.request({ rptId, startOrigin: paymentStartOrigin })
     );
 
-  const resetPayment = () => {
-    dispatch(runDeleteActivePaymentSaga());
-    dispatch(paymentInitializeState());
-  };
+  const onDuplicatedPayment = () =>
+    dispatch(
+      paymentCompletedSuccess({
+        rptId,
+        kind: "DUPLICATED"
+      })
+    );
 
   const onCancel = () => {
     dispatch(abortRunningPayment());
   };
 
+  const navigateToPaymentTransactionError = (error: TransactionSummaryError) =>
+    navigateToPaymentTransactionErrorScreen({
+      error,
+      onCancel,
+      rptId
+    });
+
   const startOrResumePayment = (
-    verifica: PaymentRequestsGetResponse,
+    paymentVerification: PaymentRequestsGetResponse,
     maybeFavoriteWallet: ReturnType<
       typeof mapStateToProps
     >["maybeFavoriteWallet"],
@@ -468,12 +408,12 @@ const mapDispatchToProps = (dispatch: Dispatch, props: OwnProps) => {
     dispatch(
       runStartOrResumePaymentActivationSaga({
         rptId,
-        verifica,
+        verifica: paymentVerification,
         onSuccess: idPayment =>
           dispatchPickPspOrConfirm(dispatch)(
             rptId,
             initialAmount,
-            verifica,
+            paymentVerification,
             idPayment,
             maybeFavoriteWallet,
             () => {
@@ -484,7 +424,7 @@ const mapDispatchToProps = (dispatch: Dispatch, props: OwnProps) => {
               navigateToPaymentPickPaymentMethodScreen({
                 rptId,
                 initialAmount,
-                verifica,
+                verifica: paymentVerification,
                 idPayment
               });
             },
@@ -493,58 +433,88 @@ const mapDispatchToProps = (dispatch: Dispatch, props: OwnProps) => {
       })
     );
 
-  const navigateToPaymentTransactionError = (
-    error: O.Option<
-      PayloadForAction<
-        | typeof paymentVerifica["failure"]
-        | typeof paymentAttiva["failure"]
-        | typeof paymentIdPolling["failure"]
-      >
-    >
-  ) =>
-    navigateToPaymentTransactionErrorScreen({
-      error,
-      onCancel,
-      rptId
-    });
-
-  const dispatchNavigateToPaymentManualDataInsertion = () =>
-    navigateToPaymentManualDataInsertion({
-      isInvalidAmount: isManualPaymentInsertion
-    });
-
-  return {
-    loadWallets: () => dispatch(fetchWalletsRequestWithExpBackoff()),
-    navigateToWalletHome: () => navigateToWalletHome(),
-    backToEntrypointPayment: () => dispatch(backToEntrypointPayment()),
-    navigateToWalletAddPaymentMethod: () =>
+  const continueWithPayment = (
+    paymentVerification: ReturnType<
+      typeof mapStateToProps
+    >["paymentVerification"],
+    maybeFavoriteWallet: ReturnType<
+      typeof mapStateToProps
+    >["maybeFavoriteWallet"],
+    hasPayableMethods: ReturnType<typeof mapStateToProps>["hasPayableMethods"]
+  ) => {
+    if (!pot.isSome(paymentVerification)) {
+      return;
+    }
+    if (hasPayableMethods) {
+      startOrResumePayment(
+        paymentVerification.value,
+        maybeFavoriteWallet,
+        hasPayableMethods
+      );
+      return;
+    }
+    alertNoPayablePaymentMethods(() =>
       navigateToWalletAddPaymentMethod({
         inPayment: O.none,
         showOnlyPayablePaymentMethods: true
-      }),
-    dispatchPaymentVerificaRequest,
+      })
+    );
+  };
+
+  const resetPayment = () => {
+    dispatch(runDeleteActivePaymentSaga());
+    dispatch(paymentInitializeState());
+  };
+
+  const startAssistanceRequest = (
+    error: TransactionSummaryError,
+    messageId: string | undefined
+  ) => {
+    resetCustomFields();
+    addTicketCustomField(zendeskCategoryId, zendeskPaymentCategory.value);
+    // Add organization fiscal code custom field
+    addTicketCustomField(
+      zendeskPaymentOrgFiscalCode,
+      rptId.organizationFiscalCode
+    );
+    if (O.isSome(error) && error.value) {
+      // Add failure custom field
+      addTicketCustomField(zendeskPaymentFailure, error.value);
+    }
+    // Add start origin custom field
+    addTicketCustomField(zendeskPaymentStartOrigin, paymentStartOrigin);
+    // Add rptId custom field
+    addTicketCustomField(zendeskPaymentNav, getCodiceAvviso(rptId));
+    appendLog(
+      JSON.stringify({
+        error,
+        messageId
+      })
+    );
+    dispatch(
+      zendeskSupportStart({
+        startingRoute: "n/a",
+        assistanceForPayment: true,
+        assistanceForCard: false,
+        assistanceForFci: false
+      })
+    );
+    dispatch(zendeskSelectedCategory(zendeskPaymentCategory));
+  };
+
+  return {
+    loadWallets: () => dispatch(fetchWalletsRequestWithExpBackoff()),
+    verifyPayment,
+    onDuplicatedPayment,
     navigateToPaymentTransactionError,
-    dispatchNavigateToPaymentManualDataInsertion,
-    startOrResumePayment,
-    isManualPaymentInsertion,
-    goBack: () => {
-      props.navigation.goBack();
-      // reset the payment state
-      dispatch(paymentInitializeState());
-    },
+    continueWithPayment,
     resetPayment,
-    onCancel,
-    onDuplicatedPayment: () =>
-      dispatch(
-        paymentCompletedSuccess({
-          rptId: props.route.params.rptId,
-          kind: "DUPLICATED"
-        })
-      )
+    backToEntrypointPayment: () => dispatch(backToEntrypointPayment()),
+    startAssistanceRequest
   };
 };
 
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(withLoadingSpinner(TransactionSummaryScreen));
+)(TransactionSummaryScreen);
