+++ ts/screens/wallet/payment/components/TransactionSummary.tsx
@@ -4,20 +4,20 @@ import * as O from "fp-ts/lib/Option";
 import React from "react";
 import { StyleSheet, View } from "react-native";
 import Placeholder from "rn-placeholder";
-import { IOColors, Icon } from "@pagopa/io-app-design-system";
-import AmountIcon from "../../../../../img/features/payments/Amount.svg";
-import CalendarIcon from "../../../../../img/features/payments/calendar.svg";
-import NoticeIcon from "../../../../../img/features/payments/Giacenza.svg";
-import { IOBadge } from "../../../../components/core/IOBadge";
-import { Body } from "../../../../components/core/typography/Body";
-import { H4 } from "../../../../components/core/typography/H4";
+import {
+  Badge,
+  Divider,
+  IOColors,
+  IOIcons,
+  Icon,
+  ListItemInfo,
+  ListItemInfoCopy
+} from "@pagopa/io-app-design-system";
 import TouchableDefaultOpacity from "../../../../components/TouchableDefaultOpacity";
-import { MultiImage } from "../../../../components/ui/MultiImage";
 import I18n from "../../../../i18n";
 import { PaymentState } from "../../../../store/reducers/wallet/payment";
 import customVariables from "../../../../theme/variables";
 import { clipboardSetStringWithFeedback } from "../../../../utils/clipboard";
-import { getLogoForOrganization } from "../../../../utils/organizations";
 import { cleanTransactionDescription } from "../../../../utils/payment";
 import {
   centsToAmount,
@@ -28,39 +28,6 @@ import { getRecepientName } from "../../../../utils/strings";
 import { format } from "../../../../utils/dates";
 
 const styles = StyleSheet.create({
-  row: {
-    flex: 1,
-    flexDirection: "row",
-    alignItems: "center"
-  },
-  vertical: {
-    flex: 1,
-    flexDirection: "column",
-    paddingVertical: customVariables.spacerHeight
-  },
-  horizontal: {
-    flex: 1,
-    flexDirection: "row",
-    paddingVertical: customVariables.spacerLargeHeight
-  },
-  icon: {
-    width: customVariables.iconSizeBase,
-    marginEnd: customVariables.spacerWidth
-  },
-  organizationIcon: {
-    width: customVariables.iconSizeBase,
-    height: customVariables.iconSizeBase
-  },
-  title: {
-    flex: 1
-  },
-  children: {
-    paddingLeft: customVariables.spacingBase
-  },
-  separator: {
-    backgroundColor: customVariables.itemSeparator,
-    height: StyleSheet.hairlineWidth
-  },
   spacer: {
     height: customVariables.spacerExtrasmallHeight
   },
@@ -68,8 +35,6 @@ const styles = StyleSheet.create({
     paddingTop: 9
   }
 });
-
-const iconProps = { color: IOColors.bluegrey };
 
 const LoadingPlaceholder = (props: { size: "full" | "half" }) => (
   <Placeholder.Box
@@ -96,63 +61,34 @@ const InfoButton = (props: { onPress: () => void }) => (
 );
 
 type RowProps = Readonly<{
-  axis: "horizontal" | "vertical";
   title: string;
-  subtitle?: string;
-  icon?: React.ReactNode;
-  placeholder?: React.ReactNode;
+  value?: string;
+  icon: IOIcons;
   isLoading?: boolean;
-  hideSeparator?: boolean;
+  placeholder?: React.ReactElement;
+  action?: React.ReactNode;
   onPress?: () => void;
 }>;
 
 export const TransactionSummaryRow = (
   props: React.PropsWithChildren<RowProps>
 ): React.ReactElement | null => {
-  if (!props.isLoading && !props.subtitle) {
-    return null;
-  }
+  const { title, value, icon, isLoading, placeholder, action } = props;
 
   return (
-    <TouchableDefaultOpacity onPress={props.onPress} accessible={false}>
-      <View>
-        <View style={styles.row}>
-          {props.icon && <View style={styles.icon}>{props.icon}</View>}
-          <View
-            style={
-              props.axis === "vertical" ? styles.vertical : styles.horizontal
-            }
-            accessible={true}
-            accessibilityLabel={`${props.title}: ${
-              props.isLoading
-                ? I18n.t("global.accessibility.activityIndicator.label")
-                : props.subtitle
-            }`}
-          >
-            <Body weight="Regular" color={"bluegrey"} style={styles.title}>
-              {props.title}
-            </Body>
-            {!props.isLoading && props.axis === "horizontal" && props.subtitle && (
-              <H4 color={"blue"} weight={"SemiBold"}>
-                {props.subtitle}
-              </H4>
-            )}
-            {!props.isLoading && props.axis === "vertical" && props.subtitle && (
-              <H4 color={"bluegreyDark"} weight={"SemiBold"}>
-                {props.subtitle}
-              </H4>
-            )}
-            {props.isLoading && (
-              <View style={styles.placeholder}>{props.placeholder}</View>
-            )}
-          </View>
-          {props.children && (
-            <View style={styles.children}>{props.children}</View>
-          )}
-        </View>
-        {props.hideSeparator !== true && <View style={styles.separator} />}
-      </View>
-    </TouchableDefaultOpacity>
+    <ListItemInfo
+      accessibilityLabel={title}
+      icon={icon}
+      label={title}
+      value={
+        !isLoading ? (
+          value
+        ) : (
+          <View style={styles.placeholder}>{placeholder}</View>
+        )
+      }
+      action={action}
+    />
   );
 };
 
@@ -173,11 +109,6 @@ export const TransactionSummary = (props: Props): React.ReactElement => {
     O.toUndefined
   );
 
-  const organizationIcon = [
-    ...getLogoForOrganization(props.organizationFiscalCode),
-    require("../../../../../img/features/payments/Institution.png")
-  ];
-
   const description = pot.toUndefined(
     pot.mapNullable(props.paymentVerification, _ =>
       cleanTransactionDescription(_.causaleVersamento)
@@ -194,7 +125,7 @@ export const TransactionSummary = (props: Props): React.ReactElement => {
     props.paymentVerification,
     pot.toOption,
     O.chainNullableK(_ => _.dueDate),
-    O.map(_ => format(_, "DD MMMM YYYY")),
+    O.map(_ => format(_, "DD/MM/YYYY")),
     O.toUndefined
   );
 
@@ -205,26 +136,29 @@ export const TransactionSummary = (props: Props): React.ReactElement => {
   const { presentPaymentInfoBottomSheet, paymentInfoBottomSheet } =
     usePaymentAmountInfoBottomSheet();
 
+  const renderAmountAction = () => {
+    if (props.isPaid && !isLoading) {
+      return <Badge text={I18n.t("messages.badge.paid")} variant="success" />;
+    } else if (!props.isPaid && !isLoading) {
+      return <InfoButton onPress={presentPaymentInfoBottomSheet} />;
+    }
+    return null;
+  };
+
   return (
     <>
       <TransactionSummaryRow
-        axis={"vertical"}
         title={I18n.t("wallet.firstTransactionSummary.recipient")}
-        subtitle={recipient}
-        icon={
-          <MultiImage
-            style={styles.organizationIcon}
-            source={organizationIcon}
-          />
-        }
-        placeholder={<LoadingPlaceholder size={"full"} />}
+        value={recipient}
+        icon="institution"
         isLoading={isLoading}
+        placeholder={<LoadingPlaceholder size={"full"} />}
       />
+      <Divider />
       <TransactionSummaryRow
-        axis={"vertical"}
         title={I18n.t("wallet.firstTransactionSummary.object")}
-        subtitle={description}
-        icon={<NoticeIcon {...iconProps} />}
+        value={description}
+        icon="notice"
         placeholder={
           <>
             <LoadingPlaceholder size={"full"} />
@@ -234,46 +168,41 @@ export const TransactionSummary = (props: Props): React.ReactElement => {
         }
         isLoading={isLoading}
       />
+      <Divider />
       <TransactionSummaryRow
-        axis={"vertical"}
         title={I18n.t("wallet.firstTransactionSummary.amount")}
-        subtitle={amount}
-        icon={<AmountIcon {...iconProps}></AmountIcon>}
+        value={amount}
+        icon="psp"
         placeholder={<LoadingPlaceholder size={"half"} />}
         isLoading={isLoading}
-      >
-        {props.isPaid && !isLoading && (
-          <IOBadge
-            small
-            text={I18n.t("messages.badge.paid")}
-            variant="solid"
-            color="aqua"
-          />
-        )}
-        {!props.isPaid && !isLoading && (
-          <InfoButton onPress={presentPaymentInfoBottomSheet} />
-        )}
-      </TransactionSummaryRow>
+        action={renderAmountAction()}
+      />
+      <Divider />
       <TransactionSummaryRow
-        axis={"vertical"}
-        title={I18n.t("wallet.firstTransactionSummary.expireDate")}
-        subtitle={dueDate}
-        icon={<CalendarIcon {...iconProps} />}
+        title={I18n.t("wallet.firstTransactionSummary.dueDate")}
+        value={dueDate}
+        icon="calendar"
         placeholder={<LoadingPlaceholder size={"half"} />}
         isLoading={isLoading}
       />
-      <TransactionSummaryRow
-        axis={"horizontal"}
-        title={I18n.t("payment.noticeCode")}
-        subtitle={formattedPaymentNoticeNumber}
+
+      <Divider />
+      <ListItemInfoCopy
+        value={formattedPaymentNoticeNumber}
+        icon="notice"
+        label={I18n.t("payment.noticeCode")}
+        accessibilityLabel={I18n.t("payment.noticeCode")}
         onPress={() =>
           clipboardSetStringWithFeedback(props.paymentNoticeNumber)
         }
       />
-      <TransactionSummaryRow
-        axis={"horizontal"}
-        title={I18n.t("wallet.firstTransactionSummary.entityCode")}
-        subtitle={props.organizationFiscalCode}
+
+      <Divider />
+      <ListItemInfoCopy
+        value={props.organizationFiscalCode}
+        icon="creditCard"
+        label={I18n.t("wallet.firstTransactionSummary.entityCode")}
+        accessibilityLabel={I18n.t("wallet.firstTransactionSummary.entityCode")}
         onPress={() =>
           clipboardSetStringWithFeedback(props.organizationFiscalCode)
         }
