--- ts/screens/wallet/payment/components/TransactionSummaryButtonConfigurations.ts
@@ -1,25 +1,0 @@
-import { IOColors } from "@pagopa/io-app-design-system";
-import { BlockButtonProps } from "../../../../components/ui/BlockButtons";
-import I18n from "../../../../i18n";
-
-export const loadingButtonProps = (): BlockButtonProps => ({
-  block: true,
-  onPress: undefined,
-  title: "",
-  disabled: true,
-  style: { backgroundColor: IOColors.greyLight, width: "100%" },
-  isLoading: true,
-  iconColor: "bluegreyDark"
-});
-
-export const continueButtonProps = (onPress: () => void): BlockButtonProps => ({
-  block: true,
-  onPress,
-  title: I18n.t("wallet.continue")
-});
-
-export const helpButtonProps = (onPress: () => void): BlockButtonProps => ({
-  block: true,
-  onPress,
-  title: I18n.t("payment.details.info.buttons.help")
-});
